Implementing deterministic shared interface resolution for structured meshes.  That is, given the starting vertex handles from neighboring processors, each processor can compute the remote handles for shared vertices, based on the remote proc's ijk space, requiring fewer shorter messages.
This commit also moves ProgOptions into MOAB proper; I think we can safely say this is a
useful addition that applications might be interested in (and it's useful for testing too).
This also moves test_prog_opt from tools into test.

Also fixing a few errors (zoltan config, refiner test file) exposed by make distcheck.

test/parallel/scdtest.cpp: re-implemented to use the netcdf reader, and to test the various
  partition schemes.
test/Makefile.am: adding test_prog_opt
src/ReorderTool.cpp: adding const to operator()
src/ScdInterface.cpp: moving lots of the partition functionality from ReadNC to here, and
  adding a global box dimension tag/member.
src/Skinner.cpp: structured mesh check depends only on ScdInterface being present, not also
  on there being a box dimension tag.
src/parallel/ParallelComm.cpp: in resolve_shared_ents, check for structured mesh
  and use deterministic resolution if present.  Had to move a few functions to public section
  to allow calling them from ScdInterface.
src/ProgOptions.cpp: moved here, from tools
src/io/ReadNC.hpp: moved partition functions to ScdInterface
src/io/ReadTetGen.hpp: added include for string
src/Makefile.am, tools/mbzoltan/mbpart.cpp, tools/Makefile.am, tools/dagmc/dagmc_preproc.cpp: 
  changes for ProgOptions move
tools/dagmc/obb_analysis.cpp: being more explicit about type to remove warning

Passes make distcheck in serial and parallel.



git-svn-id: https://svn.mcs.anl.gov/repos/ITAPS/MOAB/trunk@5108 6162379e-bd28-0410-9a7d-b7f4dcfcad3c

+++ tools/dagmc/obb_analysis.cpp
@@ -74,7 +74,7 @@ ErrorCode obbvis_create( DagMC& dag, std::vector<int> &volumes, int grid, std::s
   
   /* Compute the vertices of the visualization grid.  Calculation points are at the center 
      of each cell in this grid, so make grid+1 vertices in each direction. */
-  int numpoints = pow(grid+1,3);
+  int numpoints = pow((double)(grid+1),3);
   double* pgrid = new double[ numpoints * 3 ];
   int idx = 0;
